+++ src/App.js
@@ -9,6 +9,7 @@ import configureLogger from './configureLogger';
 import configureStore from './configureStore';
 import gatherActions from './gatherActions';
 import { destroyStore } from './mainActions';
+import { unregisterServiceWorker } from './serviceWorkerRegistration';
 
 import Root from './components/Root';
 
@@ -30,6 +31,10 @@ export default class StripesCore extends Component {
     this.epics = configureEpics(connectErrorEpic);
     this.store = configureStore(initialState, this.logger, this.epics);
     this.actionNames = gatherActions();
+
+    // unregister any zombie service workers left over from RTR work
+    // prior to disabling RTR in PR #1371
+    unregisterServiceWorker();
   }
 
   componentWillUnmount() {